Wrote the detail view
The input comes from the url (that is either directly typed or built
from user input) and then gets passed down the URLconf chain of command.

The regex captures the input as specified by the user and then passes it
along to the view callback function.

Pretty neat i must say.