Tutorial 5 on testing is done
I faced some problems where two of the tests were failing. After some
tense moments, I realized that there was a mistake on my part in the
view where i did not put a . at the end of a message. In my test case
there was a dot. So my tests were failing since the assertions were not
getting through the test runner. Lesson learned.